wsocket/requirements.txt
django==3.1.7
djangorestframework==3.12.4
websocket-client==1.1.0
requests==2.25.1